- (agent/extensible/read_config.c): ansi semantic fix for '<'.
git-svn-id: file:///home/hardaker/lib/sf-bkups/net-snmp-convert-svnrepo/trunk@415 06827809-a52a-0410-b366-d66718629ded

+++ net-snmp/agent/extensible/read_config.c
@@ -220,7 +220,7 @@ int read_config(filename, procp, numps, pprelocs, numrelocs, pppassthrus,
               cptr = skip_not_white(cptr);
               cptr = skip_white(cptr);
               if (cptr != NULL) {
-                if (strlen(cptr) < COMMUNITY_MAX_LEN) {
+                if (((int) strlen(cptr)) < COMMUNITY_MAX_LEN) {
                   copy_word(cptr,communities[i-1]);
                 } else {
                   fprintf(stderr,"snmpd.conf:  comminity %s too long\n",cptr);
